fix: move documents package to fix WalletDetails polymorphism
