+++ src/main/kotlin/it/pagopa/wallet/documents/wallets/Client.kt
@@ -1,0 +1,3 @@
+package it.pagopa.wallet.documents.wallets
+
+data class Client(val status: String)
