+++ src/main/kotlin/it/pagopa/wallet/documents/wallets/details/PayPalDetails.kt
@@ -1,4 +1,4 @@
-package it.pagopa.wallet.scheduler.documents.details
+package it.pagopa.wallet.documents.wallets.details
 
 data class PayPalDetails(val maskedEmail: String?, val pspId: String, val pspBusinessName: String) :
     WalletDetails<PayPalDetails>
