+++ src/main/kotlin/it/pagopa/wallet/documents/wallets/details/WalletDetails.kt
@@ -1,0 +1,3 @@
+package it.pagopa.wallet.documents.wallets.details
+
+interface WalletDetails<T>
