--- src/main/kotlin/it/pagopa/wallet/scheduler/documents/Client.kt
@@ -1,3 +1,0 @@
-package it.pagopa.wallet.scheduler.documents
-
-data class Client(val status: String)
