--- src/main/kotlin/it/pagopa/wallet/scheduler/documents/details/WalletDetails.kt
@@ -1,3 +1,0 @@
-package it.pagopa.wallet.scheduler.documents.details
-
-interface WalletDetails<T>
